requirements.txt
# Install with "python -m pip install -r requirements.txt".

pur==3.0.5
aiohttp==0.22.1
aiodns==1.0.1
pyyaml==3.11
cchardet==1.0.0
surt==0.3.0
robotexclusionrulesparser==1.7.0
cachetools==1.1.6
pluginbase==0.4
bottle==0.12.9
gevent==1.1.1
filemagic==1.6
coverage==4.1
tldextract==2.0.1
pylint==1.6.3
pytest==2.9.2
pytest-cov==2.3.0
